+++ src/otx/backend/native/cli/__init__.py
@@ -1,0 +1,12 @@
+"""CLI for Native backend.
+
+Note: This is temporary as the new CLI should cover all the utilities mentioned here.
+"""
+
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+
+from .utils import list_models
+
+__all__ = ["list_models"]
