+++ src/otx/backend/native/utils/adaptive_bs/__init__.py
@@ -3,6 +3,6 @@
 """API for adaptive batch size."""
 
 
-from .adaptive_bs_api import adapt_batch_size
+from .runner import adapt_batch_size
 
 __all__ = ["adapt_batch_size"]
