+++ src/otx/backend/native/utils/auto_configurator.py
@@ -1,7 +1,8 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
 """Auto-Configurator class & util functions for OTX Auto-Configuration."""
+
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
 
 from __future__ import annotations
